+++ spec/install.spec.js
@@ -88,8 +88,11 @@ describe('install', function() {
                 },
                 dependent_plugins:{}
             });
-            install('android', temp, dummyplugin, plugins_dir, {});
-            expect(spy).toHaveBeenCalledWith('results', 'Plugin "'+dummy_id+'" already installed, all good.');
+            installPromise(install('android', temp, dummyplugin, plugins_dir, {}));
+            waitsFor(function() { return done; }, 'install promise never resolved', 500);
+            runs(function() {
+                expect(spy).toHaveBeenCalledWith('results', 'Plugin "'+dummy_id+'" already installed on android.');
+            });
         });
         it('should check version if plugin has engine tag', function(){
             var spy = spyOn(semver, 'satisfies').andReturn(true);
@@ -189,7 +192,7 @@ describe('install', function() {
             });
             waitsFor(function() { return done; }, 'install promise never resolved', 500);
             runs(function() {
-                expect(actions_push.calls.length).toEqual(5);
+                expect(actions_push.calls.length).toEqual(3);
                 expect(c_a).toHaveBeenCalledWith(jasmine.any(Function), [jasmine.any(Object), path.join(plugins_dir, dummyplugin), temp, dummy_id], jasmine.any(Function), [jasmine.any(Object), temp, dummy_id]);
                 expect(proc).toHaveBeenCalled();
             });
@@ -267,6 +270,27 @@ describe('install', function() {
                 runs(function() {
                     expect(s).toHaveBeenCalledWith('D', deps_dir, { link: false, subdir: undefined, git_ref: undefined, client: 'plugman', expected_id: 'D' });
                     expect(s.calls.length).toEqual(2);
+                });
+            });
+            it('should process all dependent plugins with alternate routes to the same plugin', function() {
+                // Plugin F depends on A, C, D and E
+                runs(function () {
+                    installPromise(install('android', temp, 'F', path.join(plugins_dir, 'dependencies'), {}));
+                });
+                waitsFor(function () { return done; }, 'install promise never resolved', 500);
+                runs(function () {
+
+                    // So process should be called 5 times
+                    expect(proc.calls.length).toEqual(5);
+                });
+            });
+            it('should throw if there is a cyclic dependency', function() {
+                runs(function () {
+                    installPromise(install('android', temp, 'G', path.join(plugins_dir, 'dependencies'), {}));
+                });
+                waitsFor(function () { return done; }, 'install promise never resolved', 500);
+                runs(function () {
+                    expect(done.message).toEqual('Cyclic dependency from G to H');
                 });
             });
         });